moves version styles to separate stylesheet
